+++ CHANGELOG.md
@@ -16,6 +16,7 @@ _Not yet on NuGet..._
 * Text: Added an `OffsetX` and `OffsetY` properties for adjusting text position in pixel units (#3506) @jamaa
 * Demo: Added a demonstration window for highlight the point nearest the cursor across multiple scatter plots (#3507, #3503) @jamaa @RubensMigliore
 * Polygon: Improved automatic axis limit detection of polygons (#3501) @drphobos
+* Annotation: New plot type for adding text labels aligned to the data area which are always visible (#3510, #3356) @dlampa
 
 ## ScottPlot 5.0.22
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-03-16_